+++ js/ui/proxy-click.js
@@ -1,26 +1,27 @@
 /**
  * @module ui/proxy-click
+ * @description Used for cards and things that look like they should be clickable 
+ * even though the link in their content is the only clickable element. This way 
+ * the entire cards content doesn't need to be in a link (which isn't accessible). 
+ * 
+ * The script allows only for clicks with a duration of 250ms to avoid conflict 
+ * with a user selecting text. Works with either links or buttons because it just 
+ * uses the elements .click(). Uses data-attributes for selection by default.
  */
 
+import { ComponentInitializer } from "../utils/system.js";
 
-// Used for cards and things that look like they should be clickable even 
-// though the link in their content is the only clickable element. This way the 
-// entire cards content doesn't need to be in a link (which isn't accessible)
-// - The script allows only for clicks with a duration of 250ms to avoid 
-//   conflict with a user selecting text.
-// - Works with either links or buttons because it just uses the elements .click()
-// - Uses data-attributes for selection
-import { getName } from "../events/index.js";
-import { getDatasetOptionalJson } from "../utils/dom.js";
+/**
+ * Proxy Click Component Initializer
+ */
+export const initializer = new ComponentInitializer({ 
+  type: "proxy-click", 
+  baseAttribute: "data-ulu-proxy-click"
+});
 
-const attrs = {
-  trigger: "data-ulu-proxy-click",
-  init: "data-ulu-proxy-click-init",
-};
-
-const attrSelector = key => `[${ attrs[key] }]`;
-const attrSelectorInitial = key => `${ attrSelector(key) }:not([${ attrs.init }])`;
-
+/**
+ * Default options
+ */
 export const defaults = {
   selector: "[data-ulu-proxy-click-source]",
   selectorPreventBase: "input, select, textarea, button, a, [tabindex='-1']",
@@ -32,7 +33,7 @@ export const defaults = {
 let currentDefaults = { ...defaults };
 
 /**
- * @param {Object} options Change options used as default for dialogs, can then be overriden by data attribute settings on element
+ * @param {Object} options Change options used as default for dialogs, can then be overridden by data attribute settings on element
  */
 export function setDefaults(options) {
   currentDefaults = Object.assign({}, currentDefaults, options);
@@ -42,26 +43,39 @@ export function setDefaults(options) {
  * - This will only initialize elements once, it is safe to call on page changes
  */
 export function init() {
-  document.addEventListener(getName("pageModified"), () => setup());
-  setup();
-}
-
-export function setup(context = document) {
-  const proxies = context.querySelectorAll(attrSelectorInitial("trigger"));
-  proxies.forEach(proxy => {
-    const elOptions = getDatasetOptionalJson(proxy, "siteProxyClick");
-    const options = Object.assign({}, currentDefaults, elOptions);
-    const child = proxy.querySelector(options.selector);
-    if (child) {
-      attachHandlers(proxy, child, options);
-      proxy.setAttribute(attrs.init, "");
-    } else {
-      console.error("Unable to locate proxy click source", options.selector);
+  initializer.init({
+    withData: true,
+    events: ["pageModified"],
+    setup({ element, data, initialize }) {
+      setupProxy(element, data);
+      initialize();
     }
   });
 }
-export function attachHandlers(proxy, child, options) {
-  const { selectorPreventBase: spb, selectorPrevent: sp } = options;
+
+/**
+ * Setup a single proxy click
+ * @param {Node} proxy The container who's click should proxy the click of inner element with options.selector (defaults to [data-ulu-proxy-click-source])
+ * @param {Object} userOptions Options to override defaults
+ */
+export function setupProxy(proxy, userOptions) {
+  const options = Object.assign({}, currentDefaults, userOptions);
+  const child = proxy.querySelector(options.selector);
+  if (child) {
+    attachHandlers(proxy, child, options);
+  } else {
+    console.error("Unable to locate proxy click source", options.selector);
+  }
+}
+
+/**
+ * Main function for attaching behaviors that enable proxy click
+ * @param {Node} proxy The container who's click should proxy the click of inner element with options.selector (defaults to [data-ulu-proxy-click-source])
+ * @param {Node} child The element who is being proxied and will get clicked if the proxy is clicked (as long as not an interactive element within proxy)
+ * @param {Object} config Merged/final options object
+ */
+export function attachHandlers(proxy, child, config) {
+  const { selectorPreventBase: spb, selectorPrevent: sp } = config;
   const selectorPrevent = `${ spb }${ sp ? `, ${ sp }` : "" }`;
   let start, shouldProxy;
   proxy.addEventListener("mousedown", ({ target, timeStamp }) => {
@@ -72,7 +86,7 @@ export function attachHandlers(proxy, child, options) {
     }
   });
   proxy.addEventListener("mouseup", ({ timeStamp }) => {
-    if (shouldProxy && timeStamp - start < options.mousedownDurationPrevent) {
+    if (shouldProxy && timeStamp - start < config.mousedownDurationPrevent) {
       child.click();
     }
   });